+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+
+## 0.10.103
+
+_2023-09-13_
+
 - feat: Allow `BlingComboProps` values as `blingType` for `NewsHero` and
   `InfoHero`
 - fix: Correctly pass `ssr` prop down to sub-`Tabs`